Use an anonymous memory map as backing for the image buffer
That way, the system knows what exactly to swap and we can only
hope that the layout in memory prevents too much trashing to occour
in low-memory situations. The latter is probably not the case
and I feel that ultimately each thread could write into a Vec
which alter gets aggregated.

In any case, I'd expect the memory pressure to be reduced significantly
now.