dial back and only create one image at a time.
When seeing that one would, of course, hold multiple images in memory
with the previous signature for marginal if any performance improvements
I thought it would be best to instead just loop over themes.

On the bright side, the way threads receive their data does not
rely on owning it anymore, but instead we just make sure each thread
only accesses a single index, one at a time, over a synchronized
counter.